WIP: parse all responses from a single server together
NOT TO BE MERGED YET

This changes polling.clj to make all the requests to a single server up
front, and pass all of the results together to a single parsing
function.

The use case we're working towards is: sensu's "silencing" feature
publishes which alerts are silenced at a different endpoint to where the
alerts themselves are published (/stashes and /events respectively).  We
therefore need our alerts parsing function to have the results from
multiple web requests.

This is unfinished because a) it hasn't converted icinga, and b) it has
removed the error handling which handle-response used to provide.