Follow clojure namespace naming conventions
Fixes the contribution from #20 to follow the normal convention of using
dashes rather than underscores in namespace names.

Hyphens are preferred to underscores in namespace names (although the
filenames on disk must use underscores, due to JVM limitations [1]).

I've only checked this continues to compile, not actually tested it
against a running server.

[1] http://stackoverflow.com/questions/4420944/why-does-clojure-convert-dashes-in-names-to-underscores-in-the-filesystem