Migrate to component structure
In order to make working in the REPL more pleasant when dealing with a
web server and long running pollers, migrating to component will allow
us to manage the lifecycle of these piece more easily.

This has separated out the management of Poller instances from the
loading of configuration, which was previously not a great conflation of
concerns.

If you want to read more about the component library and application
structure go here:

https://github.com/stuartsierra/component