Ensure well ordered history of requests
As we will always trigger a request every 10 seconds, it is possible
that a request that was started after a previous could return before. We
do not want to overwrite newer data with older stale information. To
ensure that we do not, we have introduced a store of timestamps when
each key was last requested. Using this we can ensure that data is never
overwritten by a response that was started before the latest.